+++ Controllers/SeasonController.php
@@ -27,17 +27,19 @@ class SeasonController
     function LoadSeasons(){
         $seasons=$this->GetSeasons();
         $logged = $this->user_controller->CheckLoggedIn();
-        $this->view->RenderSeasons($seasons,$logged);
+        $admin = $this->user_controller->checkAdmin();
+        $this->view->RenderSeasons($seasons,$logged,$admin);
     }
     function LoadEdit($params = null)
     {
         $logged = $this->user_controller->CheckLoggedIn();
-        if ($logged) {
+        $admin = $this->user_controller->checkAdmin();
+        if ($logged && $admin) {
             $id_edit = $params[':ID'];
             $seasons = $this->model->GetSeasons();
             $season_to_edit =  $this->model->GetSeasons($id_edit);
             
-            $this->view->RenderSeasonEdit($seasons, $logged,$season_to_edit );
+            $this->view->RenderSeasonEdit($seasons, $logged,$season_to_edit,$admin);
         } else {
             $this->view->RenderError('no estas loggeado', 'logueate e intenta de nuevo');
         }
@@ -46,7 +48,8 @@ class SeasonController
     {
     if (isset($_POST['season_number_input']) && isset($_POST['chapter_count_input'])) {
         $logged = $this->user_controller->CheckLoggedIn();
-        if ($logged) {
+        $admin = $this->user_controller->checkAdmin();
+        if ($logged && $admin) {
             $id_edit = $params[':ID'];
             $this->model->UpdateSeason($_POST['season_number_input'],$_POST['chapter_count_input'],$id_edit);
             header('location:'.BASE_URL.'seasons');
@@ -58,7 +61,8 @@ class SeasonController
     function DeleteSeason($params = null)
     {
         $logged = $this->user_controller->CheckLoggedIn();
-        if ($logged) {
+        $admin = $this->user_controller->checkAdmin();
+        if ($logged && $admin) {
             $id_borrar = $params[':ID'];
             $this->chapterModel->DeleteAllChapters($id_borrar);
             $this->model->DeleteSeason($id_borrar);
@@ -71,8 +75,9 @@ class SeasonController
     {
         $logged = $this->user_controller->CheckLoggedIn();
         $seasons = $this->model->GetSeasons();
-        if ($logged) {
-            $this->view->RenderUploadSeasonModo($logged,$seasons);
+        $admin = $this->user_controller->checkAdmin();
+        if ($logged && $admin) {
+            $this->view->RenderUploadSeasonModo($logged,$seasons,$admin);
         } else {
             $this->view->RenderError('no estas loggeado', 'logueate e intenta de nuevo');
         }
@@ -80,7 +85,8 @@ class SeasonController
     function InsertSeason()
     {
         $logged = $this->user_controller->CheckLoggedIn();
-        if ($logged) {
+        $admin = $this->user_controller->checkAdmin();
+        if ($logged && $admin) {
             if (isset($_POST['season_number_input']) && isset($_POST['chapter_count_input'])) {
                 $this->model->InsertSeason($_POST['season_number_input'],$_POST['chapter_count_input']);
                 header('location:'.BASE_URL.'seasons');
